+++ variables.py
@@ -7,5 +7,5 @@ GREEN_BACK="\x1b[42;30m"
 YELLOW_BACK="\x1b[43;30m"
 RESET="\x1b[0m"
 
-#Singleton redis to prevent connection conflicts
+# Singleton redis to prevent connection conflicts
 r = redis.Redis(host='127.0.0.1', port=6379)
