+++ workers/arduino/arduino_sensor_worker.py
