+++ workers/pi/sensor_worker.py
@@ -3,6 +3,7 @@ import datetime
 import json
 import redis
 import threading
+from .worker import Worker
 import sys
 sys.path.append('..')
 from sensors.pi.float_sensor import (FloatSensor)
@@ -10,33 +11,18 @@ from sensors.pi.humidity_sensor import (HumiditySensor)
 
 import variables
 
-#r = redis.Redis(host='127.0.0.1', port=6379)
-# def clamp(n, smallest, largest): return max(smallest, min(n, largest))
 
-class PiSensorWorker():
+class PiSensorWorker(Worker):
 	def __init__(self, config, main_thread_running, system_ready):
-		#self.config = {**config, **self.config}
-		self.config = config
+		super().__init__(config, main_thread_running, system_ready)
 		self.channel = config.get('channel', 'sensors').replace(" ", "_").lower()
 		self.sleep_duration = config.get('sleep_duration', 30)
-		self.main_thread_running = main_thread_running
-		self.system_ready = system_ready
-		#Store pump event so we can shutdown pump with float readings
+
 		self.sensors = []
-		self.init_sensors()
+		self.init()
 		return
 
-	def dynamic_import(self, name):
-		#Split path of the class folder structure: {sensor name}_sensor . {SensorName}Sensor
-		components = name.split('.')
-		#Dynamically import root of component path
-		module = __import__(components[0])
-		#Get component attributes
-		for component in components[1:]:
-			module = getattr(module, component)
-		return module
-
-	def init_sensors(self):
+	def init(self):
 		for sensor in self.config['sensors']:
 			if sensor.get('type', None) is not None:
 				#Get the sensor from the sensors folder {sensor name}_sensor.{SensorName}Sensor
@@ -71,13 +57,10 @@ class PiSensorWorker():
 		return
 
 	def run(self): 
-		t = threading.Thread(target=self.work, args=())
-		t.start()
 		print('Pi Sensor Worker [' + str(len(self.sensors)) + ' Sensors]...\t\t\033[1;32m Online\033[0;0m')
-		return t
+		return super().run()
 
 	def work(self):
-
 		while self.main_thread_running.is_set():
 			if self.system_ready.is_set():
 				message = {'event':'PiSensorUpdate'}
@@ -97,8 +80,7 @@ class PiSensorWorker():
 							else:
 								pass
 								#self.pump_ready.clear()
-						
-
+					
 				print(readings)
 				message['data'] = readings
 				variables.r.publish(self.channel, json.dumps(message))
